LPS-162363 Fix styling and rename SwitchButton component
