+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/_support-overview.scss
@@ -1,5 +1,5 @@
 @import './components/SLACardsList/sla-cards-list';
 
 .cp-support-overview {
-	width: 520px;
+	max-width: 520px;
 }
