+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/_sla-cards-list.scss
@@ -1,11 +1,13 @@
 @import './components/SLACard/sla-card';
 @import './components/SLACardMessage/sla-card-message';
+@import './components/SwitchSLACardButton/switch-sla-card-button';
 
-.cp-sla-container:hover .btn {
-	display: inline-block !important;
-	padding: 0.25rem;
-	transform: translate(11.5rem, -2.5rem);
-	width: 2rem;
-	height: 2rem;
-	z-index: 100;
+.cp-sla-container {
+	&:hover {
+		.cp-switch-sla-card-button {
+			display: inline-block !important;
+			transform: translate(10.5rem, -2.5rem);
+			z-index: 100;
+		}
+	}
 }
