+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/components/SLACardMessage/SLACardMessage.js
@@ -13,7 +13,7 @@ import i18n from '../../../../../../../../../../../common/I18n';
 
 const SLACardMessage = () => (
 	<div className="bg-neutral-1 cp-sla-card-message rounded-lg">
-		<p className="px-3 py-2 text-neutral-7 text-paragraph-sm">
+		<p className="m-0 px-3 py-2 text-neutral-7 text-paragraph-sm">
 			{i18n.translate(
 				"the-project's-support-level-is-displayed-here-for-projects-with-ticketing-support"
 			)}
