+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/components/SLACardMessage/_sla-card-message.scss
@@ -1,4 +1,3 @@
 .cp-sla-card-message {
-	height: 5rem;
 	width: 12rem;
 }
