+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/components/SwitchSLACardButton/SwitchSLACardButton.test.js
