+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/components/SwitchSLACardButton/_switch-sla-card-button.scss
@@ -1,0 +1,4 @@
+.cp-switch-sla-card-button {
+	width: 2rem;
+	height: 2rem;
+}
