+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/components/SupportOverview/components/SLACardsList/components/SwitchSLACardButton/index.js
@@ -9,4 +9,4 @@
  * distribution rights of the Software.
  */
 
-export {default} from './SwitchButton';
+export {default} from './SwitchSLACardButton';
